RJMCMC.jl Library

Documentation for RJMCMC.jl

RJMCMC.RJChain — Type

RJChain

Contains information about the chain including a vector of accepted states

Fields

`n::Int32 : number of jumps in chain

`states::Vector{Any} : vector of states

`jumptypes::Vector{Int32} : vector of the attempted jumptype from the current state

`α::Vector{Float32} : vector of acceptance probabilities from proposed state

`accept::Vector{Bool} : vector of results from acceptance calculation

RJMCMC.RJMCMCStruct — Type

RJMCMCStruct

Contains information about creating the chain

Fields

burnin::Int32: number of steps for chain burn in.

`iterations::Int32' : number of steps for the returned chain

`njumptypes::Int32' : number of different jump types

`jumpprobability::Vector{Float32}' : probability mass function for jump type selection

`proposalfuns' : array of functions for proposal functions

`acceptfuns' : array of functions for acceptance functions

proposal and acceptance functions must have the signature

(proposedstate,vararg)=proposalfuns[jt](mhs,currentstate)     
α=acceptfuns[jt](mhs,currentstate,proposedstate,vararg)

where jt is the index of the jumptype, mhs is structure passed to the functions, and vararg is anything needed to be passed between related propose and accept fucntions

RJMCMC.buildchain — Method

buildchain(rjs::RJMCMCStruct,mhs,intialstate)

Generate RJMCMC chain and returns a RJChain struct.